closes #11, #5 Fixes issues with LOCALE
Breaking Change Release:
- Removed per-core option as the defaults changed and honestly I dont think they ever made sense. Without knowing the number of cores load averages are a useless metric to monitor.

Non Breaking Changes:
- moved common logic our of check-load.rb
- metrics-load.rb now use common logic
- fixed common logic for LOCALE
