+++ CHANGELOG.md
@@ -4,10 +4,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+## [3.0.0]
+### Breaking Change
+- bin/check-load.rb remove option for per_core as its useless and broken (@majormoses)
+- bin/metrics-load.rb remove option for per_core as its useless and broken (@majormoses)
+
+### Changed
+- moved common logic into its own library location (@majormoses)
+### Fixed
+- bin/metrics-load.rb now uses common logic which fixes issues with uptime LOCALE parsing (@majormoses)
+
 ### [2.0.0]
 ### Breaking Change
 - bin/check-load.rb change the default of per_core to true (@majormoses)
-
 ### Changed
 - Use the `uptime` command in bin/check-load.rb as well (@smuth4)
 - Correctly detect the number of cores on BSD (@smuth4)
@@ -43,7 +52,8 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Added
 - initial release
 
-[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-load-checks/compare/2.0.0...HEAD
+[Unreleased]: https://github.com/sensu-plugins/sensu-plugins-load-checks/compare/3.0.0...HEAD
+[3.0.0]: https://github.com/sensu-plugins/sensu-plugins-load-checks/compare/2.0.0...3.0.0
 [2.0.0]: https://github.com/sensu-plugins/sensu-plugins-load-checks/compare/1.0.0...2.0.0
 [1.0.0]: https://github.com/sensu-plugins/sensu-plugins-load-checks/compare/0.0.4...1.0.0
 [0.0.4]: https://github.com/sensu-plugins/sensu-plugins-load-checks/compare/0.0.3...0.0.4
