+++ README.md
@@ -21,11 +21,10 @@ To see the list of full options you can run:
 $ ./bin/check-load.rb --help
 Usage: ./bin/check-load.rb (options)
     -c, --crit L1,L5,L15             Load CRITICAL threshold, 1/5/15 min average
-    -p, --per-core                   Divide load average results by cpu/core count
     -w, --warn L1,L5,L15             Load WARNING threshold, 1/5/15 min average
 ```
 
-This check will only work on linux systems as it relies on `cat /proc/loadavg` and `cat /proc/cpuinfo`. The check now defaults to using the `per_core` option which will take the loadavg and divide by the number of cores. You can use `-w/-c` with a comma separated value for 1/5/15 minute thresholds.
+This check should work on linux systems and many other unix systems. The check takes the loadavg and divide by the number of cores. You can use `-w/-c` with a comma separated value for 1/5/15 minute thresholds.
 
 
 ## Installation
