+++ lib/sensu-plugins-load-checks/version.rb
@@ -1,6 +1,6 @@
 module SensuPluginsLoadChecks
   module Version
-    MAJOR = 2
+    MAJOR = 3
     MINOR = 0
     PATCH = 0
